apps/api/pyproject.toml
[project]
name = "publisher-billing-api"
version = "0.1.0"
description = "Publisher Billing Lite API"
requires-python = ">=3.12"
dependencies = [
    "alembic",
    "fastapi",
    "asyncpg",
    "sqlalchemy>=2,<3",
    "tqdm>=4.66",
    "uvicorn[standard]",
    "bcrypt>=4.0,<4.1",
    "passlib[bcrypt]>=1.7.4",
    "PyJWT>=2.8.0",
]

[dependency-groups]
dev = [
    "mypy>=1",
    "pre-commit>=3",
    "ruff>=0.6",
    "types-tqdm>=4.67.0.20250809",
    "types-passlib>=1.7.7",
]
test = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6",
    "httpx>=0.27",
    "greenlet>=3",
]

[tool.uv]
package = false

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"alembic/env.py" = ["E402"]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
warn_unused_ignores = true
warn_redundant_casts = true
no_implicit_optional = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
